Added comments to explain c*r. Included more tests.
Most of the additional tests are for testing my understanding of c*r.
c*r is difficult to understand at first so it's important to understand
the fundamentals: car and cdr. c*r builds upon car and cdr. Deciphering
c*r involves reading the 'a's and 'd's backward from right to left.